修正

+++ app/controllers/items_controller.rb
@@ -35,8 +35,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == item.user_id
+      item.destroy
+      redirect_to root_path
+    else
+      render :show
+    end
   end
 
   private
